Added Map view
1. Added a button to home screen that will display all items on a map

In this code, I'm using the Geocoder class to convert the latitude and longitude to an address and storing it in the holidayAddress instance variable when the holiday is loaded. We then use this variable when formatting the SMS message. If the latitude and longitude are null or if the address cannot be found, holidayAddress will be null.